[C10D](reland) Add GIL checker to NCCL watchdog monitor
Whenever the monitor thread kills the watchdog thread for being stuck,
we do so to save cluster time and get a faster failure signal, but we
want to know more about why it got stuck.

One possible reason for watchdog stuckness is GIL contention, which
could be ruled out or observed by making an attempt to acquire the GIL
at exit time.

If we cannot acquire the GIL within a short time window (1s) we abort
the attempt and report GIL contention, otherwise we report that GIL was
acquired successfully.

Reland: uses a function pointer to avoid destructor ordering issues on
dlclose. (Looks like the destructor for the std::function was being run later
than the libtorchpython lib was unloaded, leading to a crash).

ghstack-source-id: 4088d898f32301271b4f6d0363e93903eb2c5a68
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117309