Rename test_legacy_vmap.py TestCase names
The problem is that the dynamo_test_failures logic recognizes tests by
their TestClass.test_name. Unfortunately we have duplicate
TestClass.test_name in test_legacy_vmap and test_vmap. This PR
unduplicates them.

Something more robust would have been to include the test file name in
the dynamo_test_failures logic, but... it's a bit too late for that. We
can fix it if it becomes more of a problem in the future.

ghstack-source-id: 9e2692066dfac2f88b063f6db1a1ffbb21bb8f7b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117320