[dynamo] Refactor install_global_once, remove usages of install_global_unsafe
We split install_global_once into two APIs:
- `install_global_by_id(prefix, value) -> name`: installs a global if it hasn't
been installed yet
- `install_global(prefix, value) -> name`: always installs the global (and
  generates a unique name for it)

Then, we refactor most callsites of `install_global_unsafe` to one of
the previous. Some callsites cannot be refactored because we create the
global name first, do a lot of stuff with it, and then install it.

This fixes more test flakiness.

Test Plan:
- Existing tests; I can't reliably repro the flakiness

ghstack-source-id: eafd51389d2cfcf0313a8b426d57559babb8601a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118100