Fuse dict keys into a DICT_KEYS guard if we are guarding on all of them.
We group guards on keys that are "easily printable" and we
transform them into just one `DICT_KEYS`.

It feels a bit silly to first split the guards in the `builder` and then
regroup them, but this allows for a number of things:
1. Correctness. `DICT_KEYS` was trying to do too much before this set of
   PRs. Now `DICT_KEYS` will be reserved for things that are easy to
   print. If they are not then we fallback to the slower but correct
   getitem method. Even more so, if the builder places several guards on
   find which ones can be elided by the `DICT_KEYS` transformation,
   while the others will be kept there for correctness.
2. Can be applied to lazy keys. We can generalize `DICT_KEYS` so that it
   just checks that a subset of keys exist within the dict. We just have
   this info at guard creation, not at guard instantiation.

For now we just perform this optimisation for CONSTANT_MATCH, but we
plan to generalize this to other types in the future, once we have lazy
key support.

ghstack-source-id: dfffc0fd83a06b9a36f5b914d0071ef2cb7c03d3
Pull Request resolved: https://github.com/pytorch/pytorch/pull/118096